requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-docx==1.0.0
python-multipart==0.0.6
jinja2==3.1.2
uuid==1.30
docx2pdf==0.1.8
reportlab==4.0.4